refactor: Round 8 - split battle_actions, boss_heuristics, dungeon
Extract focused modules to reduce file sizes:
- battle_actions.py (593→402): Extract battle_boss_abilities.py (233 LOC)
- boss_heuristics.py (585→474): Extract boss_rule_utils.py (138 LOC)
- dungeon.py (551→428): Extract dungeon_bsp.py (138 LOC)

Total reduction: 1729→1304 LOC (-25%)
