+++ docs/refactoring/round8-baseline.md
@@ -1,0 +1,42 @@
+# Refactoring Round 8 - Baseline
+
+Generated: 2026-01-13
+Starting from Round 7 completion
+Target: Split remaining Python files over 500 LOC
+
+## Round 8 Targets
+
+### Python Core (non-data files over 500 LOC)
+| File | LOC | Technique |
+|------|-----|-----------|
+| battle_actions.py | 593 | Extract by action category |
+| boss_heuristics.py | 585 | Extract by boss type |
+| dungeon.py | 551 | Extract generation phases |
+
+### Deferred (data-heavy or already refactored)
+| File | LOC | Notes |
+|------|-----|-------|
+| engine.py | 797 | Already refactored in Round 4 |
+| artifacts.py | 682 | Data file - DB migration candidate |
+| story_data.py | 638 | Data file - DB migration candidate |
+| screens.py | 602 | Already refactored in Round 4 |
+| game.py | 601 | Core orchestration - tightly coupled |
+| player.py | 568 | Just refactored in Round 7 |
+
+## Refactoring Log
+
+| Date | File | Before LOC | After LOC | Changes Made |
+|------|------|------------|-----------|--------------|
+| 2026-01-13 | battle_actions.py | 593 | 402 | Extract battle_boss_abilities.py (233 LOC) |
+| 2026-01-13 | boss_heuristics.py | 585 | 474 | Extract boss_rule_utils.py (138 LOC) |
+| 2026-01-13 | dungeon.py | 551 | 428 | Extract dungeon_bsp.py (138 LOC) |
+
+## Summary
+
+Round 8 refactored 3 files:
+- **battle_actions.py**: 593 → 402 LOC (-32%)
+- **boss_heuristics.py**: 585 → 474 LOC (-19%)
+- **dungeon.py**: 551 → 428 LOC (-22%)
+
+Total reduction: 1729 → 1304 LOC (-25%)
+New modules created: 3 (battle_boss_abilities.py, boss_rule_utils.py, dungeon_bsp.py)
